move body of download_bulk_app_translations to separate function
this commit is just moving code
