+++ corehq/apps/app_manager/translations.py
@@ -1,3 +1,4 @@
+from collections import OrderedDict
 from lxml import etree
 import copy
 from openpyxl.shared.exc import InvalidFileException
@@ -189,6 +190,140 @@ def expected_bulk_app_sheet_headers(app):
                                                              + video_lang_list
             ])
     return headers
+
+
+def expected_bulk_app_sheet_rows(app):
+    def cleaned_row(row):
+        '''
+        :param row: A tuple representing a row in the spreadsheet
+        Returns a cleaned version of row with all instances of None
+        changed to ""
+        '''
+        return tuple(item if item is not None else "" for item in row)
+
+    # keys are the names of sheets, values are lists of tuples representing rows
+    rows = {"Modules_and_forms": []}
+
+    for mod_index, module in enumerate(app.get_modules()):
+        # This is duplicated logic from expected_bulk_app_sheet_headers,
+        # which I don't love.
+        module_string = "module" + str(mod_index + 1)
+
+        # Add module to the first sheet
+        row_data = cleaned_row(
+            ("Module", module_string) +
+            tuple(module.name.get(lang, "") for lang in app.langs) +
+            tuple(module.case_label.get(lang, "") for lang in app.langs) +
+            (module.media_image, module.media_audio, module.unique_id)
+        )
+        rows["Modules_and_forms"].append(row_data)
+
+        # Populate module sheet
+        rows[module_string] = []
+
+        for list_or_detail, case_properties in [
+            ("list", module.case_details.short.get_columns()),
+            ("detail", module.case_details.long.get_columns())
+        ]:
+            for detail in case_properties:
+
+                field_name = detail.field
+                if detail.format == "enum":
+                    field_name += " (ID Mapping Text)"
+                elif detail.format == "graph":
+                    field_name += " (graph)"
+
+                # Add a row for this case detail
+                rows[module_string].append(
+                    (field_name, list_or_detail) +
+                    tuple(detail.header.get(lang, "") for lang in app.langs)
+                )
+
+                # Add a row for any mapping pairs
+                if detail.format == "enum":
+                    for mapping in detail.enum:
+                        rows[module_string].append(
+                            (
+                                mapping.key + " (ID Mapping Value)",
+                                list_or_detail
+                            ) + tuple(
+                                mapping.value.get(lang, "")
+                                for lang in app.langs
+                            )
+                        )
+
+                # Add rows for graph configuration
+                if detail.format == "graph":
+                    for key, val in detail.graph_configuration.locale_specific_config.iteritems():
+                        rows[module_string].append(
+                            (
+                                key + " (graph config)",
+                                list_or_detail
+                            ) + tuple(val.get(lang, "") for lang in app.langs)
+                        )
+                    for i, annotation in enumerate(detail.graph_configuration.annotations):
+                        rows[module_string].append(
+                            (
+                                "graph annotation {}".format(i + 1),
+                                list_or_detail
+                            ) + tuple(
+                                annotation.display_text.get(lang, "")
+                                for lang in app.langs
+                            )
+                        )
+
+        for form_index, form in enumerate(module.get_forms()):
+            form_string = module_string + "_form" + str(form_index + 1)
+            xform = form.wrapped_xform()
+
+            # Add row for this form to the first sheet
+            # This next line is same logic as above :(
+            first_sheet_row = cleaned_row(
+                ("Form", form_string) +
+                tuple(form.name.get(lang, "") for lang in app.langs) +
+                tuple("" for lang in app.langs) +
+                (form.media_image, form.media_audio, form.unique_id)
+            )
+
+            # Add form to the first street
+            rows["Modules_and_forms"].append(first_sheet_row)
+
+            # Populate form sheet
+            rows[form_string] = []
+
+            itext_items = OrderedDict()
+            try:
+                nodes = xform.itext_node.findall("./{f}translation")
+            except XFormException:
+                nodes = []
+
+            for translation_node in nodes:
+                lang = translation_node.attrib['lang']
+                for text_node in translation_node.findall("./{f}text"):
+                    text_id = text_node.attrib['id']
+                    itext_items[text_id] = itext_items.get(text_id, {})
+
+                    for value_node in text_node.findall("./{f}value"):
+                        value_form = value_node.attrib.get("form", "default")
+                        value = value_node.text
+                        itext_items[text_id][(lang, value_form)] = value
+
+            for text_id, values in itext_items.iteritems():
+                row = [text_id]
+                for value_form in ["default", "audio", "image", "video"]:
+                    # Get the fallback value for this form
+                    fallback = ""
+                    for lang in app.langs:
+                        fallback = values.get((lang, value_form), fallback)
+                        if fallback:
+                            break
+                    # Populate the row
+                    for lang in app.langs:
+                        row.append(values.get((lang, value_form), fallback))
+                # Don't add empty rows:
+                if any(row[1:]):
+                    rows[form_string].append(row)
+    return rows
 
 
 def process_modules_and_forms_sheet(rows, app):
